docs(Installation): Added the pip install page and updated the github install
Added information and the commands to pip install as well as mentioned the recommended python versions and how to install in a virtual environment (both conda and venv). Also updated the github installation so that it is only two commands about cloning the directory as well as installing all the dependencies. The github installation is the recommended installation for developers so it has both sections for users and developers