docs(github-installation-gifs): Updated all the gifs filepaths in github installation page
To make the gifs load in github-pages I had to direct it through the full file path https://belowlab.github.io/drive/screencasts/. Now all the images will load if I use that as in the src part of the img tag. This solution would be better if it didn't hard code the path but for now it works.