refactor(filter.py): added a message that tells how many individuals from the phenotype file were actually found in the IBD files
Previously, there was no message telling you if individuals from the phenotype file were not found in the IBD data. The filter.py filter does filter to only grids that are in the phenotype file but before v2.7.4 there was no clear message saying X/Y number of grids from cohort were identified. Now there is a log message with the info priority meaning it will be logged by the default verbosity settings.