fix(filter.py): Couldn't read in build 38 data where the chromosome has prefix chr
There was a bug where I was only checking for a integer to identify the correct chromosome in the filter file. This step was failing if the file was build38 where the chromosome has a prefix of chr and therefore cannot be converted to an integer. I fixed this by checking for both options in the filter