fix(filter.py-and-pyproject.toml): Adjust code so that chromosomes can also use chr prefix
Sometimes chromosomes in data files are alphanumeric (chr12) instead of just numeric (12). Updated code to check for this. The code adjust the self.target_gene.chr attribute if the chromosome has the chr prefix for downstream filtering use. If neither of the chromosome ids are in the data then DRIVE still raises an ValueError