fix(filter.py): removed the chromosome check
Forgot that self.ibd_file is a IOWrapper which was causing the comparison in line 141 to fail. I removed the chromosome check. There user can make sure the files line up. I might add an error message if the pairs_df id empty that the user should make sure it is the correct chromosome