refactor(cluster.py): Changed the minimum network size to be more intuitive and allow reporting of 2 person networks
Previously, lines 112-116 in the cluster.py file reported networks that were > than the min_cluster_size. This means that if we wanted networks of size 2 we would have to change this minimum threshold to 1. This design wasn't necessarily intuitive so instead I changed it to be >=. I also change the default value of the min-network-size flag (the flag that sets this attribute) to 3 to preserve legacy behavior by default. The user shouldn't notice a difference.