test(updated-the-integration-test): Updated integration test to use simulated IBD data as test data
Found a way to simulate IBD data using msprime. This simulated data is now the input for the test cases in drive. Also updated the pytest package so it is a non dev dependency that way all individuals will have it installed. And updated syntax for integration test because it was very out of date.