feat(dendrogram): Adding the ability to generate a dendrogram from the drive networks
There were instructions that functionality has to be added to be able to generate dendrograms from the DRIVE output directly instead of using the python package ibd-dendrogram. To accomplish this DRIVE is being restructed to use subcommands. One subcommand will handle the clustering while teh other will handle the dendrogram.