feat(network_algorithm.py): moved relevant code for the network algorithm into its own folder
In an attempt to create structure and isolate code for the subcommands, cluster and dendrogram, I created two sub-directories. One is called dendrogram and the other is called network. Dendrogram will have the necessary code for the dendrogram script (although it may reuse some classes from the networking side) and network has all the DRIVE clustering scripts. I also updated the parsers a little bit so they are better structured. I also used a builder type pattern so that the appropriate function for either the dendrogram or the cluster commands is set as one of the callable args. This final change simplifies the main "drive" script so that it just calls the args.func instead of having to determine which subcommand function to call.