feat(generate_dendrogram.py): Began to create the logic to construct dendrograms from the DRIVE output files
Added logic for reading in the ids from the networks, generating indices for each IBD input, splitting the target gene like in the clustering algorithm, and filtering the ibd segments. Right now the ibd segments filtering is the same IbdFilter class from the clustering algorithm. This class needs to be changed at some point because it is tightly coupled to the formation of the graph object.