fix(testing-and-parser/__init__.py): added integration test for dendrogram functionality and fixed bug where isort messed up imports
Isort reorders teh imports in the parser/__init__.py and causes a circular import. I moved the imports back to how they were and then added the tag to skip this file from sorting. Also added another integration test that attempts to run the dendrogram function and checks if the output plot was made. Maybe there could be better logic down the line for more sophisticated test but this is okay for now