fix(pvalue.py): Fixed a bug in the carrier count value going to the binomial teset
In running DRIVE on new data, I noticed that there was a bug when after removing the proband the carrier count value being passed to the binomial test function would go to 0. This bug was introduced when we corrected how the proband was being removed in the stats. Now DRIVE returns a p-value of 0 if there are no carriers in the network or if there is only 1 carrier to account for the proband being removed. The 1 carrier in the network will still be described in the output file but the p-value will now be 0.