docs(drive.py): Added code that prints out the full command provided to DRIVE
added a logger.info statment that will print out the full command provided to DRIVE in both STDOUT and the log file if specified. This command will help with record keeping. Also updated the "examples_command_format.rst" documentation to reflect that depending on the environment you are running your command in, you might want to use more flags the the basic cluster command. This is reflected in a hint at the bottom of the page