feat(utilities.pull_samples): Added another subcommand that allows users to pull samples from networks inthe DRIVE results
One current challenge that has been mentioned in using the DRIVE results is that most tools that allow you to filter by samples require you to have a text file where each row is a sample id. DRIVE outputs a comma-separated string of samples in network because the output file is more focus on the network information as compared to individual information. This new pull-samples command allows users to now to grab the samples from a specific network and have it formatted for a tool like bcftools. At the moment it only does 1 network at a time."