feat(filter.py): Switch pandas read_csv to c-engine
Switched the pandas read_csv to c engine from the default python engine to see if this would speed up the filter.py script where chunks of the dataframe are read into memory. When profiled this section accounted for ~27% of total runtime so any speedup might be appreciated as long as it maintained the ability to read in the files.