refactor(filter.py): changed the column data types to use less memory when readin in chunks of the dataframe
To reduce memory consumption, I assigned smaller types to the IBD dataframe in teh filter.py script. The pair1 and pair2 id string became a string["pyarrow"], the start and end position became an int32 because each chromosome will never be more than 1,000,000,000 base pairs (chr2 is around 220,000,000 and it is the largest), and converted the cM length to a float32 instead of a float64. I tried a float16 but this was giving an overflow runtime warning in the unit test for the dendrogram.