docs(docs): Update docs to include more granular installation instructions
Began updating the documentation so that so it has more information about the python versions that are compatible with DRIVE. Added a statement in the landing page about the version requirements. Added a section in both the pip installation and github installation about the required python versions. Also created a FAQ page that addresses things like what python versions does python work with, what is the difference between DRIVE v1 and DRIVE v3, how to design a plugin, etc.