docs(docs/): Updated testing instructions and the deleted the automatic class detection
I updated the testing instructions to reflect that there is a specific way to run the test if you are using DRIVE v3 as compared to using versions v1 or v2. Also got rid of the autoclass detection because it doesn't add anything to the documentation site and it makes it more error prone. If someone wants to contribute to the code they can read through the codebase.