test(cmdline_parser.py): Started to merge integration test into DRIVE itself
For people not use to python, it can be confusing to use pytest to run the integration test for DRIVE. One alternative is to use PDM but then the user has to understand PDM and python's pyproject.toml syntax. Another method is to integrate the testing in a DRIVE subcommand using the pytest.main() function. This change starts that integration. We still have to figure out how to make the test data avaliable for the test using this method