feat(dendrogram_subcommand.py): Added ability to map the IDs from a network to anonymous ids
Before publication, we commonly map the ids from a DRIVE network to non-internal IDs (patient_1, patient_2, etc.). I added this functionality to the DRIVE dendrogram so that it will relabel the leaf nodes for the user. The program will also save the mapping in the sam eoutput directory as the dendrogram. This behavior can be turned on or off by passing the map-ids boolena flag to the commandline