feat(run_integration_test.py): Integrated Testing into the DRIVE utilities
One of the reviews of DRIVE focused on how the tool didn't work when they tried the test data. This error partially occurred because of a python version mismatch on their part, but it also occurred because asking academic users to be familar with pytest is probably not reasonable. It seemed like a better idea to incorporate the integration tests into a DRIVE command called test that under the hood runs the pytest.main() function. Refactored the pyproject.toml so that it includes the tests directory in the build. We now rely on using the sysconfig package to find the platlib install path for the virtualenv. This means of course that we are assuming you are installing DRIVE into a virtualenv (conda, venv, or a PDM virtualenv). The unit tests are not guarenteed to work outside of a virtualenv. If the program is being install with PDM then it is still best to run using "pdm run pytest -v tests/test_integration.py" from the main folder of the DRIVE github. This simplifies running the tests to only 2 different methods from 3 before. TODO: add new test structure to docker