docs(Testing-commands): Updated commands to show how to run integration testing based on how DRIVE was installed
With the new version of DRIVE, v3.0.2, I integrated testing into the DRIVE cli so that users wouldn't have to know how to run pytest to get the integration test to work. I updated the documentation to reflect thow the user can use this command if DRIVE was install from PYPI. I also showed how they could run the command using PDM if that was the method of install or how they could use python -m to run the tests if they installed the dependences using a conda yaml file or pyproject.toml with pip. Also updated styling so that these different options are on different tabs as oppose to the web page being a giant list