fix(run_integration_test.py): updated the script to look for integration test in multiple potential places
The integration test can be located in several places depending on how DRIVE is installed. These situations are numbered for clarity. 1) The PyPI install and the container install bundle the test in the virtualenv. 2) PDM doesn't include the test in the created virtualenv. 3) pip doesn't install the test when installing required packages in a venv or conda environment (Note this install method is only meant for those who clone the repo and are not contributing). Option 1 was working but I had to add paths to find the test in other spots such as tests/test_integration.py for options 2 & 3.