docs(README.md): updatd README to reflect new testing structure
Now that all of the test are integrated behind the 'drive utilities test' command, I updated the documentation to show how to run the commands for pip install, PDM install, github install, and docker/singularity install. This update includes commands to check the versioning and to run the integration test.