feat(network_writer.py): added ability to compress the network output file from DRIVE cluster
When run phenome-wide, the output from the clustering command of DRIVE can get relatively large (on the order of several GB). This size is not large by any means but if many DRIVE jobs (on the order of 100s) are run then the user could quick run into storage management issues. Added this flag to allow the user to compress the output file. This adds a new attribute to the Data Container API called config_options that allows for key:value pairs of information used during runtime. Also renamed to the data container from Data to RunTimeState because it is used to convey and share state of the application which includes Data but also includes other information. Removed the use of the Data_Interface because it seems unnecessary to make an interface that just describes attributes of an object and no methods. At that point, we might as well just use the data type."