feat(phenotype_descriptions_parser.py): updated code to automatically load phecode descriptions and add phecode name to min-pvalue col in the output file
Previously DRIVE required the user to upload a descriptions column if they wished to have DRIVE annotate the most enriched phecode with its name. This step seemed burdensome so I changed the code so now DRIVE stores the phecode mappings and automatically reads the files in. It will then annotate the most enriched phecode by default if the user is running enrichment for phecodes in phecode 1.2 or phecode X. Additionally this simplied the filtering output by phecode categories code because the program also loads in which phecodes belong to which category and now the user can just provide the phecode category (as long as it is spelled right) and it will filter to that group. This behavior replaces having to provide the phecode prefix