fix(phenotype_descriptions_parser.py): updated bug where some phecodes didn't have a category
The 1000+ phecodes from phecode1.2 don't have categories. This was causing the dictionary to behave a little weird. Now we can just use the term other to specify those. Also added an additionaly error check in the network_writer where if no phecodes remain after trying to filter the phecode columns to the user specified ones, then the program logs an error message and terminates.