docs(expected_plugin_structure.rst): updated comments in example plugin code to show where user should add code
The phrasing for where users should put their own code in the plugin was not quite clear. I added a statement saying which method the user should put their code into. Additionally I added a comment that describes how to figure out what the plugin name needed for the initialize method should be based on the file name