feat(network_writer.py): added able to split output into files based on phecode categories
Added the ability to break up the output file by phecode category to the program. This addition is different from the user selecting specific phecodes to write output for. The user can pass the flag "--split-phecode-categories" and DRIVE will create a file for each phecode group. Each file will have the same first 10ish columns with all of the information about the networks and the minimum p-value across all phecodes. This change should make the output more managable when DRIVE is run phenomewide for something like PheCode X. This flag should only be used if the user is running DRIVE phenomewide for either PheCode 1.2 or PheCode X