feat: Added ability to determine phenotype frequency within related individuals
It has been observed that the enrichment test sometimes finds "echo" signals during the phenomewide analysis (a signal that shows up in very unrelated spots of the genome). It was proposed that this may be caused by a bias due to the phenotype frequency being calculated in the entire cohort. This change adds the ability to only calculate the phenotype frequency in the related individuals of the cohort. This feature is in an experimental alpha stage and needs to be tested before it is pushed to main.