feat(case_file_parser.py): added way to record phenotype frequency in smaller cohort
The change to use only the individuals in the networks to determine phenotype frequencies seems to have had the opposite effect where the echo phenotypes are more enriched now. This result is unexpected so I added a flag that indicates that the user wants to record the new phenotype frequencies so we can check that. This flag will also be useful in general to see the new phenotype frequencies.