refactor(case_file_parser.py): remove the related individual adjustment in the phenotype parser
We attempted to correct for some of the echos that we were seeing in the output by adjusted the statistical calculation to only those individuals who were clustered into networks. This ended up not working because most individuals are clustered into networks (~90%) so therefore the stats were barely changing. We are trying an additional approach now where we drop out those VWF cases who are closely related from the analysis before running DRIVE. This means we can get rid of the filter function. I kept the function to record the phenotype frequencies because this still may be helpful for some users. I additionally started moving some of the more complicated type definitions to their own file so that I could create a type alias.