test(test_phecode_descriptions.py): started implementing unit tests for drive with the phecode descriptions parser
DRIVE has integration tests but no unit tests which has sometimes allowed bugs to pop up. This commit starts creating unit tests by focusing on the phecode_descriptions_parser.py file (because it is the easiest). I had to restructure the load_phenotype_descriptions function to make it easier to test. It use to be responsible for creating the path for the phecode files and gathering the files. It should raise a FileNotFoundError if i failed to find the 2 required files. Since it was making the path, it was hard to create a dummy path for testing. Now the network_algorithm.py file creates the path and passes it o the load_phenotype_descriptions function