feat(filters/loci_filters.py): started moving IBD file filters into their own module
Mulitple parts of the code base now need to filter thee IBD segments to just those regions of interest. I created a filters module that will have scripts used to handle this file filtering. There are filters version for pandas, polars (using parquet), and duckdb because we are moving to support those as inputs for performance benefits