refactor(generate_indices.py): created v2 of generate_indices function
The indices code is very verbose with the OOP heavy classes. I created an new version that relies on a namedtuple type IbdFileIndices as the type for the indices object. Depending on what ibd_file_format option the user provides, the code now creates the object with the proper indices. This function is not used in the wider codebase yet but will soon replace the other way