refactor: started to refactor the IbdFilter class in networks/filters/filter.py
Goal is to move the filter in networks/filters/filter.py to be outside of the module so that other parts of the code base are not reaching into the network module to interact with the ibd file. This first step begins that decoupling process and fixes broken unit test.