feat(drive/filters): added scripts for DuckDB filtering steps
Created a class that can properly generate the DuckDB SQL strings that will be used to filter the IBD files before going into the clustering analysis. Added a function to the ibd_file_reader.py script that uses duckdb to read int all of the IBD segments while filtering for who is in the cohort, what is the minimum segment size in the analysis, and whether the size overlapped or contained the locus of interest. Also finalized the filters in the loci_filter.py script