test(test_ibd_file_reader_benchmark.py): added a benchmark to test performance of the file reader script
added a benchmark that profiles the functions involved in reading in and preparing the IBD files to be read into a graph. This benchmark exceeds just a single function because it is import how that whole section of code is functioning as a whole. It test samples at 1k, 70k, and 250k participants