fix(cluster.py): Fixing a stale graph bug and igraph bug
A stale graph was being passed to the redo_clustering function and behavior would change based on the order that individuals were being read in. Additionally the wrong attribute was being used to read the haplotypes from the networks. This attribute was being used as a hack for buggy behavior in igraph about what metadata the vertices have. That behavior now seems to be behavior appropriately so we can just use the haplotype labels